apps/backend/requirements.txt
django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
python-dotenv==1.2.1
django-stubs==5.1.3
djangorestframework-stubs==3.15.3
pytest-django==4.10.0
factory-boy==3.3.3